+++ src/TreatmentOptimization/Analysis/plotTreatmentOptimizationJointVelocities.m
@@ -3,7 +3,7 @@
 % This function reads .sto files for experimental joint angles, and a
 % states file in the treatment optimization results directory, and an osim
 % model file. It plots joint angles for all joints specified as states for
-% GPOPS2. There is an option to plot multiple model files by passing in a 
+% GPOPS2. There is an option to plot multiple model files by passing in a
 % list of output states files in the modelDataFiles argument.
 %
 % There are 2 optional arguments for figure width and figure height. If no
@@ -40,8 +40,19 @@
 % ----------------------------------------------------------------------- %
 
 function plotTreatmentOptimizationJointVelocities(modelFileName, ...
-    trackedDataFile, modelDataFiles, figureWidth, figureHeight)
+    trackedDataFile, resultsDataFiles, varargin)
 import org.opensim.modeling.Storage
+params = getPlottingParams();
+if nargin > 3
+    options = parseVarargin(varargin);
+else
+    options = struct();
+end
+if isfield(options, "useRadians")
+    useRadians = options.useRadians;
+else
+    useRadians = 1;
+end
 model = Model(modelFileName);
 trackedDataStorage = Storage(trackedDataFile);
 [coordinateLabels, trackedDataTime, trackedData] = parseMotToComponents(...
@@ -50,90 +61,126 @@ trackedData = trackedData';
 if trackedDataTime(1) ~= 0
     trackedDataTime = trackedDataTime - trackedDataTime(1);
 end
-for i = 1 : size(trackedData, 2)
-    if model.getCoordinateSet().get(coordinateLabels(i)).getMotionType() ...
-        .toString().toCharArray()' == "Rotational"
-        trackedData(:, i) = trackedData(:, i) * 1;
-    end
-end
-for j=1:numel(modelDataFiles)
-    modeledStatesStorage = Storage(modelDataFiles(j));
-    [modeledStatesLabels, modeledStatesTime, modeledStates] = ...
-        parseMotToComponents(model, modeledStatesStorage);
-    modeledStates = modeledStates';
-    if modeledStatesTime ~= 0
-        modeledStatesTime = modeledStatesTime - modeledStatesTime(1);
-    end
-    
-    modeledVelocities{j} = modeledStates(:, size(modeledStates, 2)/2+1:end);
-    modeledVelocitiesLabels = modeledStatesLabels(size(modeledStates, 2)/2+1:end);
-    modeledVelocitiesTime{j} = modeledStatesTime;
-
-    for i = 1 : size(modeledStates(:, 1:size(modeledStates, 2)/2), 2)
-        if model.getCoordinateSet().get(modeledStatesLabels(i)).getMotionType() ...
-            .toString().toCharArray()' == "Rotational"
-            modeledVelocities{j}(:, i) = modeledVelocities{j}(:, i) * 1;
-
+trackedTimeNormalized = trackedDataTime ./ trackedDataTime(end);
+if ~useRadians
+    for i = 1 : size(trackedData, 2)
+        if model.getCoordinateSet().get(coordinateLabels(i)).getMotionType() ...
+                .toString().toCharArray()' == "Rotational"
+            trackedData(:, i) = trackedData(:, i) * 180/pi;
         end
     end
 end
-experimentalSpline = makeGcvSplineSet(trackedDataTime, ...
-    trackedData, coordinateLabels);
-trackedVelocities = evaluateGcvSplines(experimentalSpline, coordinateLabels, ...
-    trackedDataTime, 1);
-resampledExperimentalVelocities = {};
-for j = 1 : numel(modelDataFiles)
-    resampledExperimentalVelocities{j}= evaluateGcvSplines(experimentalSpline, ...
-        coordinateLabels, modeledVelocitiesTime{j}, 1);
+for j=1:numel(resultsDataFiles)
+    resultsStatesStorage = Storage(resultsDataFiles(j));
+    [resultsStatesLabels{j}, resultsStatesTime, resultsStates] = ...
+        parseMotToComponents(model, resultsStatesStorage);
+    resultsStates = resultsStates';
+    if resultsStatesTime ~= 0
+        resultsStatesTime = resultsStatesTime - resultsStatesTime(1);
+    end
+    resultsTimeNormalized{j} = resultsStatesTime ./ resultsStatesTime(end);
+    
+    resultsVelocities{j} = resultsStates(:, size(resultsStates, 2)/2+1:end);
+    resultsVelocitiesLabels{j} = resultsStatesLabels{j}(size(resultsStates, 2)/2+1:end);
+    resultsVelocitiesTime{j} = resultsStatesTime;
+    if ~useRadians
+        for i = 1 : size(resultsStates(:, 1:size(resultsStates, 2)/2), 2)
+            if model.getCoordinateSet().get(resultsStatesLabels(i)).getMotionType() ...
+                    .toString().toCharArray()' == "Rotational"
+                resultsVelocities{j}(:, i) = resultsVelocities{j}(:, i) * 180/pi;
+    
+            end
+        end
+    end
 end
-if nargin < 4
-    figureWidth = ceil(sqrt(numel(modeledVelocitiesLabels)));
-    figureHeight = ceil(numel(modeledVelocitiesLabels)/figureWidth);
-elseif nargin < 5
-    figureHeight = ceil(sqrt(numel(modeledVelocitiesLabels)));
+
+trackedDataSpline = makeGcvSplineSet(trackedDataTime, ...
+    trackedData, coordinateLabels);
+trackedVelocities = evaluateGcvSplines(trackedDataSpline, coordinateLabels, ...
+    trackedDataTime, 1);
+
+% Reorder labels
+for j = 1 : numel(resultsDataFiles)
+    [~, ~, indices] = intersect(resultsVelocitiesLabels{1}, resultsVelocitiesLabels{j}, 'stable');
+    resultsVelocities{j}(:, 1:length(indices)) = resultsVelocities{j}(:,indices);
+    resultsStatesLabels{j}(1:length(indices)) = resultsStatesLabels{j}(indices);
+end
+resultsVelocitiesLabels = resultsVelocitiesLabels{1};
+resultsStatesLabels = resultsStatesLabels{1};
+
+resampledTrackedVelocities = {};
+for j = 1 : numel(resultsDataFiles)
+    resampledTrackedVelocities{j}= evaluateGcvSplines(trackedDataSpline, ...
+        coordinateLabels, resultsVelocitiesTime{j}, 1);
+end
+if isfield(options, "figureGridSize")
+    figureWidth = options.figureGridSize(1);
+    figureHeight = options.figureGridSize(2);
+else
+    figureWidth = ceil(sqrt(numel(resultsVelocitiesLabels)));
+    figureHeight = ceil(numel(resultsVelocitiesLabels)/figureWidth);
 end
 figureSize = figureWidth * figureHeight;
-figure(Name = "Treatment Optimization Joint Velocities", ...
-    Units='normalized', ...
-    Position=[0.05 0.05 0.9 0.85])
-colors = getPlottingColors();
+figure(Name = "Joint Velocities", ...
+    Units=params.units, ...
+    Position=params.figureSize)
 subplotNumber = 1;
 figureNumber = 1;
 t = tiledlayout(figureHeight, figureWidth, ...
     TileSpacing='compact', Padding='compact');
-xlabel(t, "Percent Movement [0-100%]")
-ylabel(t, "Joint Velocity [deg/s]")
-
-for i=1:numel(modeledVelocitiesLabels)
+xlabel(t, "Percent Movement [0-100%]", ...
+    fontsize=params.axisLabelFontSize)
+if ~useRadians
+    ylabel(t, "Joint Velocity [deg/s]", ...
+        fontsize=params.axisLabelFontSize)
+else
+    ylabel(t, "Joint Velocity [rad/s]", ...
+        fontsize=params.axisLabelFontSize)
+end
+set(gcf, color=params.plotBackgroundColor)
+for i=1:numel(resultsVelocitiesLabels)
     if i > figureSize * figureNumber
         figureNumber = figureNumber + 1;
-        figure(Name="Treatment Optimization Joint Velocities", ...
-            Units='normalized', ...
-            Position=[0.05 0.05 0.9 0.85])
+        figure(Name="Joint Velocities", ...
+            Units=params.units, ...
+            Position=params.figureSize)
         t = tiledlayout(figureHeight, figureWidth, ...
             TileSpacing='Compact', Padding='Compact');
-        xlabel(t, "Percent Movement [0-100%]")
-        ylabel(t, "Joint Velocity [deg/s]")
+        xlabel(t, "Percent Movement [0-100%]", ...
+            fontsize=params.axisLabelFontSize)
+        if ~useRadians
+            ylabel(t, "Joint Velocity [deg/s]", ...
+                fontsize=params.axisLabelFontSize)
+        else
+            ylabel(t, "Joint Velocity [rad/s]", ...
+                fontsize=params.axisLabelFontSize)
+        end
+        set(gcf, color=params.plotBackgroundColor)
         subplotNumber = 1;
     end
-    coordinateIndex = find(modeledStatesLabels(i) == coordinateLabels);
+    coordinateIndex = find(resultsStatesLabels(i) == coordinateLabels);
     if ~isempty(coordinateIndex)
         nexttile(subplotNumber);
+        set(gca, ...
+            fontsize = params.tickLabelFontSize, ...
+            color=params.subplotBackgroundColor)
         hold on
-        plot(trackedDataTime/trackedDataTime(end)*100, trackedVelocities(:, coordinateIndex), ...
-            lineWidth=2, Color = colors(1))
-        for j = 1 : numel(modelDataFiles)
-            plot(modeledVelocitiesTime{j}/modeledVelocitiesTime{j}(end)*100, modeledVelocities{j}(:, i), ...
-                lineWidth=2, Color = colors(j+1));
+        plot(trackedTimeNormalized*100, trackedVelocities(:, coordinateIndex), ...
+            LineWidth=params.linewidth, ...
+            Color = params.lineColors(1))
+        for j = 1 : numel(resultsDataFiles)
+            plot(resultsTimeNormalized{j}*100, resultsVelocities{j}(:, i), ...
+            LineWidth=params.linewidth, ...
+            Color = params.lineColors(j+1));
         end
         hold off
         titleString = [sprintf("%s", strrep(coordinateLabels(coordinateIndex), "_", " "))];
-        for j = 1 : numel(modelDataFiles)
-            rmse = rms(resampledExperimentalVelocities{j}(:, coordinateIndex) - ...
-                modeledVelocities{j}(:, i));
+        for j = 1 : numel(resultsDataFiles)
+            rmse = rms(resampledTrackedVelocities{j}(:, coordinateIndex) - ...
+                resultsVelocities{j}(:, i));
             titleString(j+1) = sprintf("RMSE %d: %.4f", j, rmse);
         end
-        title(titleString)
+        title(titleString, fontsize = params.subplotTitleFontSize)
         if subplotNumber==1
             splitFileName = split(trackedDataFile, ["/", "\"]);
             for k = 1 : numel(splitFileName)
@@ -143,18 +190,18 @@ for i=1:numel(modeledVelocitiesLabels)
                     break
                 end
             end
-            for j = 1 : numel(modelDataFiles)
-                splitFileName = split(modelDataFiles(j), ["/", "\"]);
+            for j = 1 : numel(resultsDataFiles)
+                splitFileName = split(resultsDataFiles(j), ["/", "\"]);
                 legendValues(j+1) = sprintf("%s (%d)", splitFileName(1), j);
             end
-            legend(legendValues)
+            legend(legendValues, fontsize = params.legendFontSize);
         end
         xlim("tight")
         maxData = [];
         minData = [];
-        for j = 1 : numel(modelDataFiles)
-            maxData(j) = max(modeledVelocities{j}(:, i), [], "all");
-            minData(j) = min(modeledVelocities{j}(:, i), [], "all");
+        for j = 1 : numel(resultsDataFiles)
+            maxData(j) = max(resultsVelocities{j}(:, i), [], "all");
+            minData(j) = min(resultsVelocities{j}(:, i), [], "all");
         end
         maxData(j+1) = max(trackedVelocities(:, i), [], "all");
         minData(j+1) = min(trackedVelocities(:, i), [], "all");
@@ -167,5 +214,12 @@ for i=1:numel(modeledVelocitiesLabels)
         subplotNumber = subplotNumber + 1;
     end
 end
+end
 
+function options = parseVarargin(varargin)
+    options = struct();
+    varargin = varargin{1};
+    for k = 1 : 2 : numel(varargin)
+        options.(varargin{k}) = varargin{k+1};
+    end
 end