+++ src/TreatmentOptimization/IntegratedQuantitiesPreviewTool.m
@@ -1,10 +1,8 @@
 % This function is part of the NMSM Pipeline, see file for full license.
 %
-% This function calculates the difference between the experimental and
-% predicted inverse dynamic moments for the specified coordinate.
 %
-% (struct, Array of number, 2D matrix, Array of string) -> (Array of number)
-%
+% (string) -> (None)
+% Start Design Optimization from settings file to preview integrated terms 
 
 % ----------------------------------------------------------------------- %
 % The NMSM Pipeline is a toolkit for model personalization and treatment  %
@@ -14,7 +12,7 @@
 % National Institutes of Health (R01 EB030520).                           %
 %                                                                         %
 % Copyright (c) 2021 Rice University and the Authors                      %
-% Author(s): Marleny Vega                                                 %
+% Author(s): Spencer Williams                                             %
 %                                                                         %
 % Licensed under the Apache License, Version 2.0 (the "License");         %
 % you may not use this file except in compliance with the License.        %
@@ -28,20 +26,36 @@
 % permissions and limitations under the License.                          %
 % ----------------------------------------------------------------------- %
 
-function [cost, costTerm] = calcTrackingInverseDynamicLoadsIntegrand( ...
-    costTerm, inputs, time, inverseDynamicsMoments, loadName)
-defaultTimeNormalization = true;
-[time, costTerm] = normalizeTimeColumn(costTerm, inputs, time, ...
-    defaultTimeNormalization);
+function IntegratedQuantitiesPreviewTool(settingsFileName)
+settingsTree = xml2struct(settingsFileName);
+verifyVersion(settingsTree, "DesignOptimizationTool");
+[inputs, params] = parseDesignOptimizationSettingsTree(settingsTree);
+inputs = normalizeSynergyData(inputs);
+inputs = setupMuscleSynergies(inputs);
+inputs = setupTorqueControls(inputs);
+inputs = makeTreatmentOptimizationInputs(inputs, params);
+[setup, ~] = convertToGpopsInputs(inputs, params);
+displayIntegratedQuantitesPreview(setup, inputs);
+end
 
-[idMoment, costTerm] = findDataByLabels(costTerm, ...
-    inverseDynamicsMoments, inputs.inverseDynamicsMomentLabels, loadName);
-[experimentalMoment, costTerm] = findSplinedJointMomentsByLabels( ...
-    costTerm, inputs, time, inverseDynamicsMoments);
-
-scaleFactor = valueOrAlternate(costTerm, "scale_factor", 1);
-cost = (experimentalMoment * scaleFactor) - idMoment;
-
-[cost, costTerm] = applyTermTimeRanges(cost, costTerm, time);
-cost = normalizeCostByFinalTime(costTerm, inputs, time, cost);
+function displayIntegratedQuantitesPreview(setup, inputs)
+integral = setup.guess.phase.integral;
+index = length(integral);
+metabolic = valueOrAlternate(inputs, 'calculateMetabolicCost', false);
+braking = valueOrAlternate(inputs, 'calculateBrakingImpulse', false);
+propulsive = valueOrAlternate(inputs, 'calculatePropulsiveImpulse', false);
+if propulsive
+    disp("Propulsive impulses: " + ...
+        integral(index - length(inputs.contactSurfaces) + 1 : index));
+    index = index - length(inputs.contactSurfaces);
+end
+if braking
+    disp("Braking impulses: " + ...
+        integral(index - length(inputs.contactSurfaces) + 1 : index));
+    index = index - length(inputs.contactSurfaces);
+end
+if metabolic
+    disp("Total metabolic cost: " + integral(index));
+    index = index - 1;
+end
 end