+++ src/TreatmentOptimization/TermHelpers/findSynergyIndexByLabel.m
@@ -1,10 +1,9 @@
 % This function is part of the NMSM Pipeline, see file for full license.
 %
-% This function calculates the difference between the experimental and
-% predicted inverse dynamic moments for the specified coordinate.
+% (struct, Array of double, Array of string, Array of string) -> 
+% (Array of number, struct)
 %
-% (struct, Array of number, 2D matrix, Array of string) -> (Array of number)
-%
+% Finds the index of a synergy name, saving an index for future calls.
 
 % ----------------------------------------------------------------------- %
 % The NMSM Pipeline is a toolkit for model personalization and treatment  %
@@ -14,7 +13,7 @@
 % National Institutes of Health (R01 EB030520).                           %
 %                                                                         %
 % Copyright (c) 2021 Rice University and the Authors                      %
-% Author(s): Marleny Vega                                                 %
+% Author(s): Spencer Williams                                             %
 %                                                                         %
 % Licensed under the Apache License, Version 2.0 (the "License");         %
 % you may not use this file except in compliance with the License.        %
@@ -28,20 +27,30 @@
 % permissions and limitations under the License.                          %
 % ----------------------------------------------------------------------- %
 
-function [cost, costTerm] = calcTrackingInverseDynamicLoadsIntegrand( ...
-    costTerm, inputs, time, inverseDynamicsMoments, loadName)
-defaultTimeNormalization = true;
-[time, costTerm] = normalizeTimeColumn(costTerm, inputs, time, ...
-    defaultTimeNormalization);
+function [index, term] = findSynergyIndexByLabel(term, inputs, ...
+    synergyName)
+if isfield(term, 'internalSynergyIndex')
+    index = term.internalSynergyIndex;
+else
+    index = -1;
+    nameParts = split(synergyName, "_");
+    assert(length(nameParts) > 1 && ~isnan(str2double(nameParts(end))), ...
+        "Synergy names are referenced in terms as " + ...
+        "'<group name>_<index>', such as 'RightLeg_2'.");
+    synergyGroupName = join(nameParts(1:end-1), "_");
+    synergyNumber = str2double(nameParts(end));
 
-[idMoment, costTerm] = findDataByLabels(costTerm, ...
-    inverseDynamicsMoments, inputs.inverseDynamicsMomentLabels, loadName);
-[experimentalMoment, costTerm] = findSplinedJointMomentsByLabels( ...
-    costTerm, inputs, time, inverseDynamicsMoments);
+    counter = 0;
+    for i = 1 : length(inputs.synergyGroups)
+        if strcmp(inputs.synergyGroups{i}.muscleGroupName, ...
+                synergyGroupName)
+            index = counter + synergyNumber;
+            term.internalSynergyIndex = index;
+            break;
+        end
+        counter = counter + inputs.synergyGroups{i}.numSynergies;
+    end
 
-scaleFactor = valueOrAlternate(costTerm, "scale_factor", 1);
-cost = (experimentalMoment * scaleFactor) - idMoment;
-
-[cost, costTerm] = applyTermTimeRanges(cost, costTerm, time);
-cost = normalizeCostByFinalTime(costTerm, inputs, time, cost);
+    assert(index > 0, "Unable to find synergy " + synergyName);
+end
 end